+++ service-base/src/main/java/org/eclipse/hono/service/http/AbstractHttpEndpoint.java
@@ -16,29 +16,23 @@ package org.eclipse.hono.service.http;
 import java.net.HttpURLConnection;
 import java.util.EnumSet;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 
 import org.eclipse.hono.client.ClientErrorException;
 import org.eclipse.hono.config.ServiceConfigProperties;
 import org.eclipse.hono.service.AbstractEndpoint;
-import org.eclipse.hono.service.management.OperationResult;
-import org.eclipse.hono.service.management.Result;
 import org.eclipse.hono.tracing.TracingHelper;
 import org.eclipse.hono.util.Constants;
-import org.eclipse.hono.util.RegistryManagementConstants;
 import org.eclipse.hono.util.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 
 import io.opentracing.Span;
 import io.opentracing.tag.Tags;
-import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpMethod;
-import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.MIMEHeader;
@@ -290,80 +284,5 @@ public abstract class AbstractHttpEndpoint<T extends ServiceConfigProperties> ex
                 .allowedHeader(HttpHeaders.AUTHORIZATION.toString())
                 .allowedHeader(HttpHeaders.IF_MATCH.toString())
                 .exposedHeader(HttpHeaders.ETAG.toString());
-    }
-
-    /**
-     * Writes a response based on generic result.
-     * <p>
-     * The behavior is as follows:
-     * <ol>
-     * <li>Set the status code on the response.</li>
-     * <li>If the status code represents an error condition (i.e. the code is &gt;= 400),
-     * then the JSON object passed in the result is written to the response body.</li>
-     * <li>If the result is created (the code is = 201), the JSON object is written to the response body and the given custom handler is
-     * invoked (if not {@code null}).</li>
-     * <li>Sets the status of the tracing span and finishes it.</li>
-     * <li>Ends a response.</li>
-     * </ol>
-     *
-     * @param ctx The routing context of the request.
-     * @param result The generic result of the operation.
-     * @param customHandler An (optional) handler for post processing successful HTTP response, e.g. to set any additional HTTP
-     *                      headers. The handler <em>must not</em> write to response body. May be {@code null}.
-     * @param span The active OpenTracing span for this operation. The status of the response is logged and span is finished.
-     */
-    protected final void writeResponse(final RoutingContext ctx, final Result<?> result, final Handler<HttpServerResponse> customHandler, final Span span) {
-        final int status = result.getStatus();
-        final HttpServerResponse response = ctx.response();
-        response.setStatusCode(status);
-        if (status >= 400) {
-            HttpUtils.setResponseBody(response, JsonObject.mapFrom(result.getPayload()));
-        } else if (status == HttpURLConnection.HTTP_CREATED) {
-            if (customHandler != null) {
-                customHandler.handle(response);
-            }
-            HttpUtils.setResponseBody(response, JsonObject.mapFrom(result.getPayload()));
-        }
-        Tags.HTTP_STATUS.set(span, status);
-        span.finish();
-        response.end();
-    }
-
-    /**
-     * Writes a response based on operation result (including the resource version).
-     * <p>
-     * Sets ETAG header and then calls {@link #writeResponse}
-     *
-     * @param ctx The routing context of the request.
-     * @param result The operation result of the operation.
-     * @param customHandler An (optional) handler for post processing successful HTTP response, e.g. to set any additional HTTP
-     *                      headers. The handler <em>must not</em> write to response body. May be {@code null}.
-     * @param span The active OpenTracing span for this operation. The status of the response is logged and span is finished.
-     */
-    protected final void writeOperationResponse(final RoutingContext ctx, final OperationResult<?> result, final Handler<HttpServerResponse> customHandler, final Span span) {
-        result.getResourceVersion().ifPresent(v -> ctx.response().putHeader(HttpHeaders.ETAG, v));
-        writeResponse(ctx, result, customHandler, span);
-    }
-
-    /**
-     * Gets the payload from a request body.
-     * <p>
-     * The returned JSON object contains the given payload (if not {@code null}).
-     * If the given payload does not contain an <em>enabled</em> property, then
-     * it is added with value {@code true} to the returned object.
-     *
-     * @param payload The payload from the request message.
-     * @return The payload (never {@code null}).
-     */
-    protected final JsonObject getRequestPayload(final JsonObject payload) {
-
-        return Optional.ofNullable(payload).map(pl -> {
-            final Object obj = pl.getValue(RegistryManagementConstants.FIELD_ENABLED);
-            if (obj instanceof Boolean) {
-                return pl;
-            } else {
-                return pl.copy().put(RegistryManagementConstants.FIELD_ENABLED, Boolean.TRUE);
-            }
-        }).orElse(new JsonObject().put(RegistryManagementConstants.FIELD_ENABLED, Boolean.TRUE));
     }
 }